✨ Create class SmtpStubServer
For better instanciation and better testing
