+++ src/index.ts
@@ -1,3 +1,1 @@
-export { getEmails } from './main';
-export { launchSmtpStubServer } from './smtp-server';
-export { launchWebServer } from './web-server';
+export { SmtpStubBuilder, SmtpStubServer } from './smtp-stub-server';
